add usersControllers directory with atomic controllers for async requests
